requirements.txt
openai
python-dotenv
typer
rich
